cambio tipo variable idCatMateria en inicios

+++ database/migrations/2025_03_25_164328_create_inicios_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('inicios', function (Blueprint $table) {
             $table->id('idInicio');
             $table->string('folio_preregistro')->unique();
-            $table->integer('idCatMateria');
+            $table->unsignedBigInteger('idCatMateria');
             $table->integer('idCatVia');
             $table->dateTime('fechaCreada');
             //$table->binary('archivo');
